// src/interfaces/hiragana-char.ts
export type HiraganaChar =
  | 'あ'
  | 'い'
  | 'う'
  | 'え'
  | 'お'
  | 'か'
  | 'き'
  | 'く'
  | 'け'
  | 'こ'
  | 'さ'
  | 'し'
  | 'す'
  | 'せ'
  | 'そ'
  | 'た'
  | 'ち'
  | 'つ'
  | 'て'
  | 'と'
  | 'な'
  | 'に'
  | 'ぬ'
  | 'ね'
  | 'の'
  | 'は'
  | 'ひ'
  | 'ふ'
  | 'へ'
  | 'ほ'
  | 'ま'
  | 'み'
  | 'む'
  | 'め'
  | 'も'
  | 'や'
  | 'ゆ'
  | 'よ'
  | 'ら'
  | 'り'
  | 'る'
  | 'れ'
  | 'ろ'
  | 'わ'
  | 'ゐ'
  | 'ゑ'
  | 'を'
  | 'ん';